Add separate route for goals
