+++ portfolioserver/app.py
@@ -6,8 +6,10 @@ def create_app():
 
     from . import portfolio
     from . import schemes
+    from . import goals
 
     app.include_router(portfolio.router)
     app.include_router(schemes.router)
+    app.include_router(goals.router)
 
     return app
